Modify most of the issues in the comments.
Aside from the "save by DTO" one.

Me and Ionut (I think) agreed that it is ok to pass only the list of hours and the number.
Working with a list of LocalDateTime is really tricky and this was the way to bypass it, since the DTO would require said list.